Move lib/signal to lib/conn/signal
