+++ lib/conn/signal/handle.cc
