+++ lib/numerical/Makefile.am
@@ -5,4 +5,4 @@ lib_LTLIBRARIES = libnumrk4.la
 libnumintegration_a_SOURCES = integration.cc
 libnumfunc_a_SOURCES = func.cc
 libnumrk4_la_SOURCES = rk4.cc
-libnumrk4_la_LIBADD = ../signal/libsignalhandle.la
+libnumrk4_la_LIBADD = ../conn/signal/libsignalhandle.la
